requirements.txt
Django
djangorestframework
psycopg[binary]
django-cors-headers
drf-spectacular
django-filter
Pillow
python-decouple